Added: New dangerous permission com.termux.permission.TERMUX_PLUGIN, signature permission com.termux.permission.TERMUX_SIGNATURE, PluginService bound service for plugins.
The methods for PluginService are defined in AIDL in the new module plugin-aidl. The module can be imported by plugins to get the method definitions. The TERMUX_PLUGIN permission is needed to bind to the PluginService, and external apps have to be enabled in the config.

AIDL methods:
setCallbackBinder and setCallbackService: Used to get a Binder for the PluginService to identify each plugin and get notified when it dies.
runTask: Runs a command as a Termux task in the background if the caller also has the RUN_COMMAND permission.
listenOnSocketFile: Creates a server socket in the app directory for the plugins (TermuxConstants#TERMUX_APPS_DIR_PATH/<package name>) and returns the file descriptor, which can be used for a LocalServerSocket.
openFile: Opens a file in the app directory for the plugin and returns the file descriptor.